Ensure address details are redacted
This will ensure address fields are also redacted upon customer request.

+++ lib/tasks/confidentiality.rake
@@ -1,3 +1,4 @@
+# rubocop:disable BlockLength
 namespace :confidentiality do
   desc 'Process a confidentiality request (REFERENCE=x)'
   task redact: :environment do
@@ -11,7 +12,13 @@ namespace :confidentiality do
       phone: '000000000',
       memorable_word: 'redacted',
       date_of_birth: '1950-01-01',
-      additional_info: 'redacted'
+      additional_info: 'redacted',
+      address_line_one: 'redacted',
+      address_line_two: 'redacted',
+      address_line_three: 'redacted',
+      town: 'redacted',
+      county: 'redacted',
+      postcode: 'redacted'
     }
 
     ActiveRecord::Base.transaction do
@@ -19,10 +26,12 @@ namespace :confidentiality do
       booking_request.update!(attributes)
       booking_request.activities.where(type: 'AuditActivity').delete_all
 
-      puts 'Redacting appointment...'
-      appointment.without_auditing do
-        appointment.update!(attributes)
-        appointment.audits.delete_all
+      if appointment
+        puts 'Redacting appointment...'
+        appointment.without_auditing do
+          appointment.update!(attributes)
+          appointment.audits.delete_all
+        end
       end
 
       puts 'Done!'
